backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.2.0
opencv-python>=4.8.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
pydantic>=2.5.0
aiofiles>=23.2.1
google-cloud-vision>=3.4.0